[C-6119] branded designer UI tweaks
