+++ packages/react-brand-designer/src/brand-designer/brand-designer.tsx
@@ -1,9 +1,8 @@
 import React, { FC } from "react";
 import styled, { CSSObject } from "styled-components";
 import { BrandDesignerHeader } from "./designer-header";
-import { BrandPreview } from "./brand-preview";
-import { BrandControls } from "./brand-controls";
 import { BrandDesignerProps } from "../types";
+import { BrandSidebar } from "./brand-sidebar";
 
 export const BrandDesigner: FC<BrandDesignerProps> = (opts) => {
   const { config, onChange, onSave } = opts;
@@ -19,12 +18,7 @@ export const BrandDesigner: FC<BrandDesignerProps> = (opts) => {
   return (
     <BrandDesignerContainer>
       <BrandDesignerHeader {...headerOpts} />
-      <HeaderSeparator />
-      <BrandDesignerBody>
-        <BrandControls config={config} onChange={onChange} />
-        <PreviewSeparator />
-        <BrandPreview config={config} />
-      </BrandDesignerBody>
+      <BrandSidebar config={config} onChange={onChange} />
     </BrandDesignerContainer>
   );
 };
@@ -38,36 +32,5 @@ const BrandDesignerContainer = styled.div(
     borderRadius: "12px",
     overflow: "hidden",
     backgroundColor: "#fff",
-  })
-);
-
-const BrandDesignerBody = styled.div(
-  (): CSSObject => ({
-    display: "flex",
-    flexDirection: "row",
-    flexBasis: "368px",
-  })
-);
-
-const HeaderSeparator = styled.div(
-  (): CSSObject => ({
-    position: "absolute",
-    width: "766px",
-    height: "0px",
-    left: "32px",
-    top: "160px",
-    border: "1px solid #E1E5E9",
-  })
-);
-
-const PreviewSeparator = styled.div(
-  (): CSSObject => ({
-    position: "absolute",
-    width: "334px",
-    height: "0px",
-    left: "115px",
-    top: "345px",
-    border: "1px solid #E1E5E9",
-    transform: "rotate(90deg)",
   })
 );
